Fix httpd::python.
While writing acceptance tests for puppet-httpd we noticed that the
manifest of httpd::python is not working on Ubuntu Trusty.

The httpd::python manifest was using a non-declared
$apache::params::python_package instead of
$apache::params::mod_python_package

The change made for fixing httpd::python was similarly made upstream at
puppetlabs/apache (commit 622a8804d7c928177135247bd06ab697ccfcfb15).
We tried to cherry-pick the change to maintain attribution, but it did
not cleanly apply because the context lines around the change are
different (we use http_mod and upstream uses a2mod), as well as the
module name that the variable lives.

Change-Id: I4db9298aba7da90a017527b2fb29dcd17af28efc
Co-Authored-By: Danilo Ramalho <dramalho@thoughtworks.com>