Remove unnecessary beaker tests
It is not necessary to check every file and package in the puppet
functional tests. It can be assumed that if puppet finished corretly
then it was also able to successfully find and install packages and
apply configs. We leave the service checks, however, because it is
possible for services to die due to misconfiguration after puppet starts
them.

Change-Id: I97682cb9d956803c3ab1be42fe44c0437c3d5f02